# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Building API
  description: Sistema de controle inteligente de ar-condicionado
  version: 1.0.0
paths:
...
  /api/v1/sensors/data:
    post:
      tags:
        - Sensors
      summary: Ingerir leitura de sensor IoT
      operationId: ingest_sensor_data_api_v1_sensors_data_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorDataIngest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorDataResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - HTTPBearer: []
  /api/v1/sensors:
    get:
      tags:
        - Sensors
      summary: Listar sensores registrados
      operationId: list_sensors_api_v1_sensors_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: tipo
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/SensorTipo'
              - type: 'null'
            title: Tipo
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/sensors/{sensor_id}':
    get:
      tags:
        - Sensors
      summary: Detalhes de um sensor
      operationId: get_sensor_api_v1_sensors__sensor_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: string
            title: Sensor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/sensors/{sensor_id}/data':
    get:
      tags:
        - Sensors
      summary: Histórico de leituras
      operationId: get_sensor_history_api_v1_sensors__sensor_id__data_get
      security:
        - HTTPBearer: []
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: string
            title: Sensor Id
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(1h|24h|7d|30d)$
            default: 24h
            title: Period
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: anomalies_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
            title: Anomalies Only
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: >-
                  Response Get Sensor History Api V1 Sensors  Sensor Id  Data
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/sensors/{sensor_id}/latest':
    get:
      tags:
        - Sensors
      summary: Última leitura de um sensor
      operationId: get_sensor_latest_api_v1_sensors__sensor_id__latest_get
      security:
        - HTTPBearer: []
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: string
            title: Sensor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorLatestResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/users:
    get:
      tags:
        - Users
      summary: Listar usuários (admin)
      operationId: list_users_api_v1_users_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Users Api V1 Users Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Users
      summary: Criar usuário (admin)
      operationId: create_user_api_v1_users_post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/me:
    get:
      tags:
        - Users
      summary: Perfil do usuário autenticado
      operationId: get_me_api_v1_users_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
      security:
        - HTTPBearer: []
  '/api/v1/users/{user_id}':
    get:
      tags:
        - Users
      summary: Detalhes de um usuário (admin)
      operationId: get_user_api_v1_users__user_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Users
      summary: Atualizar usuário (admin)
      operationId: update_user_api_v1_users__user_id__patch
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Users
      summary: Desativar usuário (admin)
      operationId: delete_user_api_v1_users__user_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HateoasLink:
      properties:
        rel:
          type: string
          title: Rel
        href:
          type: string
          title: Href
        method:
          type: string
          title: Method
          default: GET
      type: object
      required:
        - rel
        - href
      title: HateoasLink
...
    PaginationMeta:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        total_pages:
          type: integer
          title: Total Pages
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - total
        - page
        - size
        - total_pages
      title: PaginationMeta
...
    SensorDataIngest:
      properties:
        sensor_id:
          type: string
          pattern: '^sensor-(temperature|humidity|presence)-\d{4}$'
          title: Sensor Id
        tipo:
          $ref: '#/components/schemas/SensorTipo'
        valor:
          type: number
          title: Valor
        tick:
          type: integer
          minimum: 0
          title: Tick
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
        - sensor_id
        - tipo
        - valor
        - tick
        - timestamp
      title: SensorDataIngest
    SensorDataResponse:
      properties:
        id:
          type: integer
          title: Id
        sensor_id:
          type: string
          title: Sensor Id
        tipo:
          $ref: '#/components/schemas/SensorTipo'
        valor:
          type: number
          title: Valor
        tick:
          type: integer
          title: Tick
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        is_anomaly:
          type: boolean
          title: Is Anomaly
        received_at:
          type: string
          format: date-time
          title: Received At
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - sensor_id
        - tipo
        - valor
        - tick
        - timestamp
        - is_anomaly
        - received_at
      title: SensorDataResponse
    SensorLatestResponse:
      properties:
        sensor_id:
          type: string
          title: Sensor Id
        tipo:
          $ref: '#/components/schemas/SensorTipo'
        valor:
          type: number
          title: Valor
        tick:
          type: integer
          title: Tick
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        is_anomaly:
          type: boolean
          title: Is Anomaly
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - sensor_id
        - tipo
        - valor
        - tick
        - timestamp
        - is_anomaly
      title: SensorLatestResponse
    SensorListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/SensorResponse'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
        - data
        - meta
      title: SensorListResponse
    SensorResponse:
      properties:
        id:
          type: string
          title: Id
        tipo:
          $ref: '#/components/schemas/SensorTipo'
        status:
          $ref: '#/components/schemas/SensorStatus'
        last_seen:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Seen
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - tipo
        - status
      title: SensorResponse
    SensorStatus:
      type: string
      enum:
        - online
        - offline
        - maintenance
      title: SensorStatus
    SensorTipo:
      type: string
      enum:
        - temperature
        - humidity
        - presence
      title: SensorTipo
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          minLength: 6
          title: Password
        role:
          allOf:
            - $ref: '#/components/schemas/UserRole'
          default: visualizador
      type: object
      required:
        - email
        - password
      title: UserCreate
    UserResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - email
        - role
        - is_active
        - created_at
      title: UserResponse
    UserRole:
      type: string
      enum:
        - admin
        - operador
        - visualizador
      title: UserRole
    UserUpdate:
      properties:
        role:
          anyOf:
            - $ref: '#/components/schemas/UserRole'
            - type: 'null'
        is_active:
          anyOf:
            - type: boolean
            - type: 'null'
          title: Is Active
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...